pyproject.toml
[project]
name = "local-news-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "python-dotenv>=1.1.0",
]
